requirements.txt
streamlit>=1.39.0
requests>=2.31.0
google-generativeai>=0.8.0
google-cloud-texttospeech>=2.16.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.1
pillow>=10.0.0
